% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_helpers.R
\name{add_all_indir_and_tot}
\alias{add_all_indir_and_tot}
\title{Add all possible indirect and total effects to the `lavaan` formula}
\usage{
add_all_indir_and_tot(lav.formula, lab.sep = "__", ngroups = 1)
}
\arguments{
\item{lav.formula}{`lavaan` formula}

\item{lab.sep}{How labels should be separated}

\item{ngroups}{Nuber of groups for multiple group data}
}
\description{
Add all possible indirect and total effects to the `lavaan` formula
}
