% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schwartz_circle.R
\name{add_circle}
\alias{add_circle}
\title{A helper function to draw Schwartz circle}
\usage{
add_circle(r, ...)
}
\arguments{
\item{r}{a radius.}

\item{...}{optional, passed to \code{geom_path}.}
}
\description{
See \code{\link{schwartz_circle}}
}
