% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_helpers.R
\name{add_custom_covs}
\alias{add_custom_covs}
\title{Append lavaan syntax with group-specific covariances}
\usage{
add_custom_covs(model, group, data, cov, focal.groups)
}
\arguments{
\item{model}{character, lavaan syntax model}

\item{group}{character, grouping variable}

\item{data}{data frame}

\item{cov}{character, covariance to add,e.g. "variable1 ~~ variable2"}

\item{focal.groups}{Character vector for the groups to add the cov.}
}
\description{
Append lavaan syntax with group-specific covariances
}
\examples{
cov.model <-  "swb =~ swb1 + swb2 swb3 + swb4 + swb5"
cov.model.custom.covs <-
cov.model \%>\%
  addCustomCovs("country", dat1,
                "swb1 ~~ swb3", c("China", "Indonesia")) \%>\%

  addCustomCovs("country.f", dat1,
                "swb2 ~~ swb4", c("Israel"))

}
