% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schwartz_circle.R
\name{add_label}
\alias{add_label}
\title{A helper function to draw Schwartz circle}
\usage{
add_label(r, angle.r, label, pos = 1/2, ...)
}
\arguments{
\item{r}{radius.}

\item{angle.r}{vector of angles to which labels are located.}

\item{label}{vector of character labels.}

\item{pos}{position between center of circle and circle, by default is 1/2, which is middle.}

\item{...}{optional, passed to \code{geom_text}; important is \code{angle} which accepts vectors and rotates the labels..}
}
\description{
See \code{\link{schwartz_circle}}
}
