% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schwartz_circle.R
\name{add_radius}
\alias{add_radius}
\title{A helper function to draw Schwartz circle}
\usage{
add_radius(angles.r1, r1, r0 = 0, ...)
}
\arguments{
\item{angles.r1}{a vector of angles.}

\item{r1}{length of line.}

\item{r0}{where the line begins, 0 by default}

\item{...}{optional, passed to \code{geom_segment}.}
}
\description{
See \code{\link{schwartz_circle}}
}
