% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_helpers.R
\name{aggr_and_merge}
\alias{aggr_and_merge}
\alias{aggr.and.merge}
\title{Creates country level variables by aggregating individual level variables}
\usage{
aggr_and_merge(ind_data, ind_var, group_var, FUN = "mean", suffix = NULL)
}
\arguments{
\item{ind_data}{A dataset with individual-level data.}

\item{ind_var}{Variable to be aggregated.}

\item{FUN}{Aggregation function, mean by default.}

\item{suffix}{Character string, what should be added to the aggregated names of variables.}

\item{country_var}{Grouping variable.}
}
\description{
Creates country level variables by aggregating individual level variables
}
