% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MplusHelpers.R
\name{checkMplusModel}
\alias{checkMplusModel}
\title{Check NDP related to residuals and correlations of models read by 'readModels'}
\usage{
checkMplusModel(model)
}
\arguments{
\item{model}{A single model read by MplusAutomation::readModels}
}
\value{
Returns a single data frame with unstandardized parameters.
}
\description{
Prints if any variance is negative oand whether there are correlations above 1 (OUTPUT: stdyx; is required in the Mplus input)
}
