% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{compare_cor}
\alias{compare_cor}
\title{Compare two correlation matrices}
\usage{
compare_cor(corr.X, corr.Y)
}
\arguments{
\item{corr.X}{First correlation matrix, e.g. produced by `cor` function}

\item{corr.Y}{Secon correlation matrix, e.g. produced by `cor` function}
}
\description{
Produces data frame of correlations with stars standing for significance
}
