% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{cor_by_country}
\alias{cor_by_country}
\title{Within-group correlations}
\usage{
cor_by_country(var1, var2, group, data, plot = TRUE, labs = TRUE,
  use = "pairwise", ...)
}
\arguments{
\item{var1}{String. Name of variable to correlate.}

\item{var2}{String. Name of variable to correlate.}

\item{group}{Grouping variable.}

\item{data}{Data.framer containing var1, var2, and group.}

\item{plot}{Logical. Should the plot be created?}

\item{labs}{Logical. Should value labels be used?

Prints correlations in console, and plots in a graph.}
}
\description{
Within-group correlations
}
