% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_helpers.R
\name{cor_within}
\alias{cor_within}
\title{Within-group correlations}
\usage{
cor_within(
  var1,
  var2,
  group,
  data,
  plot = TRUE,
  labs = TRUE,
  use = "pairwise",
  highlight.group = NA
)
}
\arguments{
\item{var1}{String. Name of variable to correlate.}

\item{var2}{String. Name of variable to correlate.}

\item{group}{Grouping variable.}

\item{data}{Data.framer containing var1, var2, and group.}

\item{plot}{Logical. Should the plot be created?}

\item{labs}{Logical. Should value labels be used?}

\item{highlight.group}{Characater. What group should be highlighted on graph?}
}
\description{
Prints correlations in console, and optionally plots in a graph.
}
