% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MplusHelpers.R
\name{diff_test_mlr}
\alias{diff_test_mlr}
\title{Loglikelihood test for MLR estimator}
\usage{
diff_test_mlr(L0, c0, p0, L1, c1, p1)
}
\arguments{
\item{L0}{loglikelihood values of reduced model}

\item{c0}{scaling correction factors of reduced model}

\item{p0}{number of parameters, should be p0 < p1}

\item{L1}{loglikelihood values of full model}

\item{c1}{scaling correction factors of full model}

\item{p1}{number of parameters, should be p0 < p1}
}
\description{
Adapted from http://www.statmodel.com/chidiff.shtml
}
