% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MplusHelpers.R
\name{diff_test_mlr_manual}
\alias{diff_test_mlr_manual}
\title{Loglikelihood test for MLR estimator - wrapper}
\usage{
diff_test_mlr_manual(reduced, full)
}
\arguments{
\item{reduced}{reduced model read by \code{\link[MplusAutomation]{readModels}}}

\item{full}{full model read by \code{\link[MplusAutomation]{readModels}}}
}
\description{
Convenience wrapper for \code{\link[LittleHelpers]{diff_test_mlr}}
}
