% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MplusHelpers.R
\name{diffTestMLR_}
\alias{diffTestMLR_}
\title{Loglikelihood test for MLR estimator - wrapper}
\usage{
diffTestMLR_(reduced, full)
}
\arguments{
\item{reduced}{reduced model read by \code{\link[MplusAutomation]{readModels}}}

\item{full}{full model read by \code{\link[MplusAutomation]{readModels}}}
}
\description{
Convenience wrapper for \code{\link[LittleHelpers]{diffTestMLR}}
}
