% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ess.R
\name{download_ess}
\alias{download_ess}
\title{Downaload ESS data directly from website to R object with labels}
\usage{
download_ess(round, country = "all", user)
}
\arguments{
\item{round}{Number of round - from 1 to 8.}

\item{country}{Vector of country \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{iso2c} codes, or "all".}

\item{user}{Your email which you used to register in ESS website.}
}
\description{
Downaload ESS data directly from website to R object with labels
}
\details{
Don't put more than one country or more than one round - it won't work. This function will expire when ESS updates its data versions, but it happens about twice a year and can be fixed manually.

\strong{Motivation.} \code{ess} package is tuned up for Stata users and sometimes do not get all the labels; it can only download one country data at a time; when it downloads several rounds, you get a list of data instead of integrated dataset.
}
\examples{
\donttest{
ESS2 <- download_ess(round=2, country="all", "mymail@gmail.com")
ESS6.Russia <- download_ess(round=6, country="RU", "mymail@gmail.com")
}

}
\seealso{
\link{label_book}
}
