% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_osf_component.R
\name{download_osf_component}
\alias{download_osf_component}
\title{Download data folder from OSF}
\usage{
download_osf_component(
  download_folders,
  osf_node_components,
  osf_node_project,
  data_directory = "local_data/osf_pull"
)
}
\arguments{
\item{osf_node_project}{An osf_project object representing the OSF project.}

\item{data_directory}{The directory where data will be downloaded. Default is "local_data/osf_pull".}

\item{download_folder}{A character vector specifying the folder(s) to download data into.}

\item{osf_node_component}{A character vector specifying the corresponding node(s) for each component.}
}
\value{
No explicit return value. Data will be downloaded to the specified directory.
}
\description{
This function downloads a complete data folder from OSF. It allows you to specify multiple folders and nodes.
}
\examples{
\dontrun{
download_folder <- c("base")
osf_node_component <- c("Base data")
osf_node_project <- osfr::osf_retrieve_node("Your OSF project")
download_osf_data(download_folder, osf_node_component, osf_node_project)
}
}
