% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_osf_folder.R
\name{download_osf_folder}
\alias{download_osf_folder}
\title{Download data folder from OSF}
\usage{
download_osf_folder(
  osf_folders,
  osf_folders_nodes,
  osf_project_node,
  data_directory = "local_data/osf_pull"
)
}
\arguments{
\item{osf_folders}{A character vector specifying the folder(s) to download data from.}

\item{osf_folders_nodes}{A character vector specifying the corresponding node(s) for each folder.}

\item{osf_project_node}{An osf_project object representing the OSF project.}

\item{data_directory}{The directory where data will be downloaded. Default is "local_data/osf_pull".}
}
\value{
No explicit return value. Data will be downloaded to the specified directory.
}
\description{
This function downloads a complete data folder from OSF. It allows you to specify multiple folders and nodes.
}
\examples{
\dontrun{
osf_folders <- c("base")
osf_folders_nodes <- c("Base data")
osf_project_node <- osfr::osf_retrieve_node("Your OSF project")
download_osf_data(osf_folders, osf_folders_nodes, osf_project_node)
}
}
