% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MplusHelpers.R
\name{eachParamPSRMplus}
\alias{eachParamPSRMplus}
\title{Various versions of PSR}
\usage{
eachParamPSRMplus(parameters, id.parameter, iterations.range = NULL)
}
\arguments{
\item{parameters}{Array typically taken from plus-produced gh5 file in gh5$bayesian_data$parameters_autocorr   or from any other, where first dimension is parameters, second dimension is iterations, and third dimension is chains}

\item{id.parameter}{Integer id of parameter}

\item{iterations.range}{Range of iterations to use. All available are used by default (NULL).}
}
\description{
Various versions of PSR
}
