% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_values.R
\name{ess_values}
\alias{ess_values}
\title{Compute basic values indices as measured in European Social Survey}
\usage{
ess_values(data, v2 = TRUE, v4 = TRUE, v10 = TRUE, v21 = FALSE,
<<<<<<< HEAD
  center = TRUE, abbr = FALSE, suffix = "")
=======
  center = TRUE, abbr = FALSE, suffix = NULL)
>>>>>>> 3ba7ee8f1e8c681e900d55fc66ae9042bcc8c4f5
}
\arguments{
\item{data}{ESS dataset of any other dataset with the same 21 value indicator variable names}

\item{v2}{Two higher value dimensions}

\item{v4}{Four higher order values}

\item{v10}{Compute 10 value indices}

\item{v21}{Add 21 raw items}

\item{center}{Apply within-individual centering?}

\item{abbr}{Logical. Use abbreviated names, e.g. SE, ST, CO..}

\item{suffix}{NULL of character. What is added to the names of centered indices.}
}
\value{
Returns a data with appended variables.
}
\description{
Computed in line with Shalom Schwartz's instructions on the ESS EduNet website.
}
