% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation_of_binary_classifiers.R
\name{evaBin}
\alias{evaBin}
\title{Evaluation of binary classifiers}
\usage{
evaBin(p = NULL, n = NULL, tp = NULL, tn = NULL, fp = NULL,
  fn = NULL, s = NULL, s.est = NULL)
}
\arguments{
\item{p}{Numeric, the number of positives.}

\item{n}{Numeric, the number of negatives.}

\item{tp}{Numeric, the number of true positives.}

\item{tn}{Numeric, the number of true negatives.}

\item{fp}{Numeric, the number of false positives.}

\item{fn}{Numeric, the number of false negatives.}

\item{s}{Numeric, true vector with values \code{TRUE} and \code{FALSE}
or \code{1}s and \code{0}s.}

\item{s.est}{Numeric, extimated vector with values \code{TRUE} and \code{FALSE}
or \code{1}s and \code{0}s.}
}
\value{
Returns a list with the following measures:
 \item{True positive (TP)}
 \item{False positive (FP)}
 \item{True negative (TN)}
 \item{False negative (FN)}
 \item{Precision}
 \item{F1-Score (F1)}
 \item{Accuracy (ACC)}
 \item{Balanced accuracy (BACC)}
 \item{Matthews correlation coefficient(MCC)}
 \item{True postive rate (TPR)}
 \item{True negative rate (TNR)}
 \item{False positive rate (FPR)}
 \item{False negative rate (FNR)}
 \item{False desovery rate (FDR)}
}
\description{
\code{evaBin} returns different measures for the performance of binary classifiers.
}
\examples{
evaBin(7, 13, 2, 10)

}
