% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_helpers.R
\name{explained.variance.merMod}
\alias{explained.variance.merMod}
\alias{explained_variance,}
\alias{explained_variance.merMod}
\title{Computes pseudo-R2 by subtracting residual variances}
\usage{
explained.variance.merMod(base.model, tested.model)
}
\arguments{
\item{base.model}{Base model}

\item{tested.model}{Test model

Doesn't make use of Satorra-Bentler correction (should give this option in future).}
}
\description{
Computes pseudo-R2 by subtracting residual variances
}
