% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MplusHelpers.R
\name{extract_mlefa}
\alias{extract_mlefa}
\title{Extracts fit indices of the Multilevel EFA (Mplus output)}
\usage{
extract_mlefa(mplus.file)
}
\arguments{
\item{output}{Output file path.}
}
\value{
Sorted data frame with differing n of factors at each level and fit statistics. By default it extracts CFI, RMSEA, RMSEA.CI.LO, RMSEA.CI.HI, SRMR.w, SRMR.b, SABIC, AIC, Chi.sq, Chi.df, Chi.p, Chi.scale.
}
\description{
Extracts summary statistics from the Mplus output file.
}
