% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fill_tri}
\alias{fill_tri}
\title{Make a matrix full form lower or upper triangle}
\usage{
fill_tri(x, lower = T)
}
\arguments{
\item{x}{Square matrix with missing on either triangle}

\item{lower}{Logical, if the `x` has data on lower triangle.}
}
\description{
Make a matrix full form lower or upper triangle
}
