% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MplusHelpers.R
\name{gammaHat.mplus}
\alias{gammaHat.mplus}
\title{Computes Gamma fit index for SEM models fit in Mplus}
\usage{
gammaHat.mplus(mplus.model)
}
\arguments{
\item{mplus.model}{Mplus model read by \link[MplusAutomation]{readModels} function.}
}
\description{
Computes Gamma fit index for SEM models fit in Mplus
}
\details{
Adapted from \link[semTools]{moreFitIndices}
}
