% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_net}
\alias{get_net}
\title{Make a quick network graph}
\usage{
get_net(
  df = NULL,
  cv = NULL,
  n = NULL,
  title = "Model",
  legend = "none",
  gamma = 0.3,
  layout = "fr",
  start = "glasso",
  stepwise = T,
  seed = 1,
  group.id = NA
)
}
\arguments{
\item{df}{Data frame}

\item{cv}{Correlation matrix (if df is NULL)}

\item{title}{Title of the plot}

\item{legend}{Legend position}

\item{gamma}{Gamma parameter for the model selection function, see `ggmModSelect`}

\item{layout}{Default is "fr". Network layout. See more at `ggraph::ggraph`}

\item{start}{Default is "glasso", see `ggmModSelect`}

\item{stepwise}{Default is TRUE, see `ggmModSelect`}

\item{seed}{Attempt to make plots replicable.}

\item{group.id}{Vector of regex expressions to apply to variable names and use it to group variables.}
}
\description{
Make a quick network graph
}
\details{
Can be used within `mutate()` and passed to `knitr::kable` with parameter `kable.options=list(escape=FALSE)`
}
