% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MplusHelpers.R
\name{get_psr}
\alias{get_psr}
\title{Get PSR from an Mplus output file}
\usage{
get_psr(path, last = TRUE, plot = FALSE)
}
\arguments{
\item{path}{Path to an Mplus output file or a directory with several output files}

\item{last}{Logical. If TRUE (default), returns only the last PSR value, if FALSE, returns a data frame with all PSR values.}

\item{plot}{Logical. If TRUE, plots the PSR values across iterations. Default is FALSE.}
}
\value{
A numeric value (if last=TRUE) or a data frame (if last=FALSE) with PSR values across iterations.
}
\description{
Get PSR from an Mplus output file
}
