% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MplusHelpers.R
\name{getParamsMplus}
\alias{getParamsMplus}
\title{Extracts parameters from Bayesian models produces by Mplus where 'readModels' cannot help}
\usage{
getParamsMplus(file)
}
\arguments{
\item{file}{Name of the Mplus .out file.}
}
\value{
A data.frame with all parameters.
}
\description{
Extracts parameters from Bayesian models produces by Mplus where 'readModels' cannot help. Currently for latent class models only.
}
