% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_helpers.R
\name{good_table}
\alias{good_table}
\title{Function producing a good table for many ML regressions #####}
\usage{
good_table(models, fit.stats = c("ICC", "random", "random.p", "R2", "fit",
  "LRT", "REML", "VIF"), mod.names = "", show.viewer = TRUE, ...)
}
\arguments{
\item{models}{List of the fitted lmer objects}

\item{fit.stats}{What fit statistics to compute/extract and show? Possible options are "ICC", "random", "R2s", "fit", "LRT", "REML", "VIF".}

\item{mod.names}{Vector of the same length as models list, giving names to each model.}

\item{show.viewer}{Logical. Whether the resulting table should be shown in the RStudio viewer. If FALSE then the file "good_table_output.html" is saved to working directory.}

\item{...}{Arguments passed to stargazer.}
}
\description{
Function producing a good table for many ML regressions #####
}
\note{
Issues to implement:
\itemize{
\item Might be veeery slow (something to work on)
\item Add row indication of non-convergence
}
}
\examples{
#m1<- lmer(HE ~ age + gender + (1|country), ess.data.d, weights=dweight)
#m2<- lmer(HE ~ age + gender + (1+age + gender|country), ess.data.d, weights=dweight)
#good_table(list(m1, m2))

# Another option
#f1 <- as.formula(HE ~ age + gender + (1|country))
#f2 <- as.formula(HE ~ age + gender + (1|country))
#good_table(lapply(c(f1, f2), function(x) lmer(x, ess.data.d, weights=dweight)))

}
