% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_helpers.R
\name{grand_center}
\alias{grand_center}
\title{Grand mean centering of one or more variables}
\usage{
grand_center(variables, prefix = "gm.")
}
\description{
Grand mean centering of one or more variables
}
