% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotInSet.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Not In Set operator (opposite to \%in\%)}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
A vector of length \code{x} with \code{TRUE}/\code{FALSE}.
}
\description{
Not In Set operator (opposite to \%in\%)
}
\examples{
x <- c("a", 2, 3)
y <- c("b", 2, 3, 4)
x \%!in\% y
}
