% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_means_ci}
\alias{graph_means_ci}
\title{Graph means w/CIs}
\usage{
graph_means_ci(
  x,
  group,
  highlight.group = NA,
  codes = c("none", "print", "caption"),
  type = c("means", "ridges", "heat"),
  use.labels = T,
  data
)
}
\arguments{
\item{x}{Variable to aggregate.}

\item{group}{Group variable. If length is 1, assumes it's a variable name in the `data`}

\item{highlight.group}{Character vector of groups make bold.}

\item{codes}{"print" or "caption"}

\item{type}{"means", "ridges", "heat"}

\item{use.labels}{Whether or not attempt to extract value labels. TRUE by default.}

\item{data}{Optional, used only if `X` or `group` have length 1.}
}
\description{
Returns ggplot of means by group with 95% confidence intervals. Modifiable with standard ggplot geoms, scales, themes, etc.
}
\examples{
d<-data.frame(v=1:100, group=rep(1:2, 50))
graph_means_ci(d$v, d$group)
with(d, graph_means_ci(v, group))
}
