% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_helpers.R
\name{group_center}
\alias{group_center}
\title{Group-centering of one or more variables}
\usage{
group_center(variables, group, data, prefix = "g.")
}
\arguments{
\item{variables}{Character vector of variables names}

\item{group}{Character of length 1,name of grouping variable}

\item{data}{Data frame}

\item{prefix}{Character, added to the befinning of centered variable name.}
}
\value{
Returns an original data frame binded with the new centered variables
}
\description{
Group-centering of one or more variables
}
