% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_book.R
\name{lab_to_fac}
\alias{lab_to_fac}
\title{Convert labelled variable to a factor with corresponding values}
\usage{
lab_to_fac(var.labelled, print = F)
}
\arguments{
\item{var.labelled}{Labelled variable or tibble with a single variable}
}
\description{
Convert labelled variable to a factor with corresponding values
}
