% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_book.R
\name{label_book}
\alias{label_book}
\alias{label_table}
\title{Create a codebook, table of labels to explore variables and values in your data}
\usage{
label_book(df, max.vals = 25, vars = "all", view = TRUE, translate = F)
}
\arguments{
\item{df}{data.frame, usually an object obtained by reading SPSS datafile with  \code{\link[haven]{read_sav}} or \code{\link[foreign]{read.spss}}}

\item{max.vals}{integer, how many value labels per each variable shoud be listed in the table, default is 25}

\item{vars}{can be integer, character, or range of integers or characters. Variables indexes or names for getting subsets of label book.}

\item{view}{logical, whether the result should be shown in the RStudio viewer pane. Default is TRUE. If FALSE, html file named 'label_book_output.html' is saved in your working directory.}

\item{translate}{Logical. Use \code{\link[gtranslate]{translate}} to auto-translate value and variable labels to English. Default is FALSE.}
}
\description{
Create a codebook, table of labels to explore variables and values in your data
}
\examples{
## Example
\dontrun{
## 1.Read the data
#ess8<- haven::read_sav("ESS8e01.sav")

## 2. Use the function
#label_book(ess8, max.vals=11, vars=45:50)
}
}
