% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_helpers.R
\name{lav_compare}
\alias{lav_compare}
\title{Combine fit measures form several models and compare}
\usage{
lav_compare(..., what = c("cfi", "tli", "rmsea", "srmr", "bic", "df"))
}
\arguments{
\item{...}{lavaan fitted objects}

\item{what}{Character vector of fit indices as given in \pkg{lavaan} `fitMeasures()`}
}
\description{
Combine fit measures form several models and compare
}
\details{
Extracts select fit measures from lavaan objects and lists it in the tables.
}
\examples{
lav_compare(fit1, fit2, fit3, what = c("cfi.scaled", "rmsea.scaled") )


}
