% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_helpers.R
\name{lav_to_draw}
\alias{lav_to_draw}
\title{Convert lavaan to draw.io diagram
Disfunctional at the moment!}
\usage{
lav_to_draw(m, output = c("text", "file"))
}
\arguments{
\item{m}{fitted lavaan object or a lavaan syntax.}

\item{output}{Format of the output.}
}
\description{
Convert lavaan to draw.io diagram
Disfunctional at the moment!
}
\details{
Produce csv-like or mermaid code that should be pasted to draw.io
The cool thing about it is that you can manually adjust the result --  a feature unavailable in any known diagram producers for lavaan so far.
}
