% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_helpers.R
\name{lavTestScore_clean}
\alias{lavTestScore_clean}
\title{Get more comprehensible output from lavTestScore}
\usage{
lavTestScore_clean(lavaan.fit, ...)
}
\arguments{
\item{lavaan.fit}{Model fitted with lavaan}

\item{...}{Arguments passed to lavTestScore}
}
\description{
Simply applies \code{\link{lavTestScore}} and attached group labels and parameter names to the output.
}
