% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MplusHelpers.R
\name{manipulate_mplus}
\alias{manipulate_mplus}
\title{manipulate mplus file}
\usage{
manipulate_mplus(inp, target, replacement, run = T, read = T, keep.files = T)
}
\arguments{
\item{inp}{Input file path.}

\item{target}{Character. Block of Mplus code to replace.}

\item{replacement}{Character. Mplus code to replace the target. Can be a vector - in this case, multiple outputs will be created.}

\item{run}{Whether new input files should be executed.  When FALSE, returns a vector of paths to the new input files.}

\item{read}{Logical, whether read the models in. When FALSE, returns the list of paths to output files.}

\item{keep.files}{Logical. Whether input and output files shuld be kept on disk. When FALSE, all Mplus files are deleted and read in wish readModels.}
}
\value{
Sorted data frame with differing n of factors at each level and fit statistics. By default it extracts CFI, RMSEA, RMSEA.CI.LO, RMSEA.CI.HI, SRMR.w, SRMR.b, SABIC, AIC, Chi.sq, Chi.df, Chi.p, Chi.scale.
}
\description{
Replaces part of the Mplus code, runs it, and reads it in. Facilitates repeated running with a small change in input. Good for testing the effects of particular analyses.
}
