% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{match_named_vectors}
\alias{match_named_vectors}
\title{Match named vectors to a data.frame}
\usage{
match_named_vectors(...)
}
\arguments{
\item{...}{Named vectors with same of partly different names. OR a list of named vectors.}
}
\value{
The function returns a data.frame with vectors stacked as columns.
}
\description{
Match named vectors to a data.frame
}
