% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{mean_se_lower_upper}
\alias{mean_se_lower_upper}
\title{Means and SEs}
\usage{
mean_se_lower_upper(x)
}
\arguments{
\item{x}{Variable for which mean, se, lower and upper bounds of confidence interval are returned.}
}
\value{
Row with 4 columns (Mean, SE, Lower, Upper)
}
\description{
Means and SEs
}
