% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_helpers.R
\name{measurementInvariance}
\alias{measurementInvariance}
\title{Run three models of invariance and compare them}
\usage{
measurementInvariance(...)
}
\arguments{
\item{...}{any \link[lavaan]{cfa} arguments except for 'group.equal'}
}
\description{
This mimicking the deprecated semTools::measurementInvariance()
}
\seealso{
lavaan::cfa
}
