% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MplusHelpers.R
\name{measurementInvarianceMplus}
\alias{measurementInvarianceMplus}
\title{Facilitates setting up a code and running measurement invariance test in Mplus}
\usage{
measurementInvarianceMplus(model, group, data, Mplus_com = "mplus")
}
\arguments{
\item{model}{Model in Mplus format, e.g. F1 BY one two three;}

\item{group}{Character,  group variable name}

\item{data}{Data}

\item{Mplus_com}{A system call to Mplus, usually just "mplus" (default)}
}
\description{
Facilitates setting up a code and running measurement invariance test in Mplus
}
\details{
Creates three files in the working directory: mplus_temp.tab, mplus_temp.inp, and mplus_temp.out. The output is stored in the latter.
}
\examples{
# measurementInvarianceMplus(model = "F1 BY F114 F115 F116 F117;
#                                     F2 BY F118 F119  F120  F121 F122  F123;",
# group = "S003",
# data = wevs.subset,
# Mplus_com = "wine mplus"
# )
}
