% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{na_tab}
\alias{na_tab}
\title{Prints the number and percent of NAs in a dataframe}
\usage{
na_tab(data, viewer = T)
}
\arguments{
\item{data}{Data.frame}

\item{viewer}{If `TRUE``, show in Rstudio viewer, if FALSE returns a data.frame}
}
\description{
Prints the number and percent of NAs in a dataframe
}
