% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf_setup_dbd.R
\name{osf_setup_dbd}
\alias{osf_setup_dbd}
\title{OSF Setup for DBD}
\usage{
osf_setup_dbd(
  semester = "OSF folder name of the semster",
  session = "OSF folder name of the session"
)
}
\arguments{
\item{semester}{OsF folder name of the semester (character string; e.g. "2022_ws").}

\item{session}{OSF folder name of the session (character string; e.g. "04-api_access-twitter").}
}
\value{
List with string parameteres
}
\description{
This is a custom function to create a data import pipeline for the Digital behavioral data seminar. It creates the local folders for saving the data downloaded from an osf repository. ´For the package to work correctly, the following parameters must be defined in the R Environemnt: OSF_PAT (Personal Access Token for OSF) and OSF_PATH (URL of the OSF project)
}
\examples{
osf_setup_dbd("2022_ws", "01-kick_off")
}
