% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plef}
\alias{plef}
\title{Simple interaction plot with nice defaults}
\usage{
plef(modl, eff.id = 1, ...)
}
\arguments{
\item{modl}{Model fitted with lm or lmer.}

\item{eff.id}{Order number of interaction term, integer.}

\item{...}{Arguments passed to effect}
}
\description{
Simple interaction plot with nice defaults
}
