% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_helpers.R
\name{plot_latent_means}
\alias{plot_latent_means}
\title{Plot latent means}
\usage{
plot_latent_means(fit)
}
\arguments{
\item{fit}{lavaan object containing an MGCFA with latent means}
}
\value{
Returns a list of ggplot objects
}
\description{
Plot latent means
}
