% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_helpers.R
\name{potential_interactions}
\alias{potential_interactions}
\title{Poterntial interactions}
\usage{
potential_interactions(random.terms, group.level.terms, lmer.fit,
  measure = c("correlation", "interaction"))
}
\arguments{
\item{random.terms}{Individual predictors  that should be made random and associated with predictors (mediators) at group level}

\item{group.level.terms}{Predictors (mediators) at the group level}

\item{lmer.fit}{Model fitted with lmer.}

\item{measure}{Character. correlation or interaction}
}
\description{
Refits lmer model with given 'random terms' and correlates the predicted random effects with aggregated at group level variable mentioned at 'group.level.terms'
}
