% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_helpers.R
\name{potential_interactions_ind}
\alias{potential_interactions_ind}
\title{Potential individual-level interactions for lmer model}
\usage{
potential_interactions_ind(variables, modelfit)
}
\arguments{
\item{variables}{Character list of variable names in the model which are already fixed effects.}

\item{modelfit}{lmer model}
}
\value{
Returns a data frame of class "LHinteractions" supplied with a corresponding print method.
}
\description{
Takes a list of variables and computes models with all possible interactions between them, one by one.
}
