% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_helpers.R
\name{print.LHinteractions}
\alias{print.LHinteractions}
\title{Method for a nice printing of LHinteractions class}
\usage{
\method{print}{LHinteractions}(x)
}
\description{
Method for a nice printing of LHinteractions class
}
