% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_helpers.R
\name{prune_boot}
\alias{prune_boot}
\title{Network pruning}
\usage{
prune_boot(net, boot, thresh = NA)
}
\arguments{
\item{net}{`bootnetResult` network}

\item{boot}{`bootnet` object, should be bootstrapped network from `net` argument.}

\item{thesh}{cutoff level (absolute) below which the edges are removed. By default, NA, keeps all the significant edges.}
}
\description{
Removes all the edges that proved to be either non-significant or smaller than the specified threshold. Returns `bootnetResult` object.
}
