% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{pvalue_to_stars}
\alias{pvalue_to_stars}
\title{Convert numbers to stars}
\usage{
pvalue_to_stars(vector.of.pvalues, na = "")
}
\arguments{
\item{vector.of.pvalues}{self-explaining argument}

\item{na}{return this if `est` is `NA`}
}
\description{
Convert numbers to stars
}
