% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{random_plot}
\alias{random_plot}
\title{Plot random effects and interactions for lmer objects}
\usage{
random_plot(lmer.fit, optional.names = NA, facets = FALSE, print = T)
}
\arguments{
\item{lmer.fit}{A mer object}

\item{optional.names}{A named vector of random effects, where names indicate terms and values are used as axis x labels in correspomding plot. The length  should be equal to a number of random effects. Usually it includes "(Intercepts)" and names of variables whose effects are made random.}

\item{facets}{Logical. Should the random effects be plotted in facets or as a series of single plots?}

\item{print}{If the plots should be actually printed, valid only if facets = F}
}
\value{
Returns one or several ggplots. In case one plot is returned it can be appended with `theme`, `geom_`, etc.
}
\description{
Takes object produced by lme4::lmer() function, extracts random effects and interactions, and returns ggplot2-based plots to examine.
}
\examples{
\dontrun{
random_plot(lmr,
   optional.names=c(`(Intercepts)`="Intercepts",
   gndr ="Female", s.Age = "Effects of age"),
   facets=TRUE)
   }
}
