% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_by_table.R
\name{replace_by_table}
\alias{replace_by_table}
\title{Replace by table}
\usage{
replace_by_table(variable = all.data$cntry, lookup.table = recodes2,
  matching.column = "cntry", replacement.column = "WDIcode",
  save.levels.order = T)
}
\arguments{
\item{variable}{Variable to recode.}

\item{lookup.table}{Data frame with (at least) two columns named in line with the following arguments: `matching.column` and `replacement.column`}

\item{matching.column}{Character. Lookup table's column of containing (pattern) values simiar to ones in `variable`}

\item{replacement.column}{Character. Lookup table's column containing substitution values.}

\item{save.levels.order}{Logical. If the level order should follow the matching.column's levels order.}
}
\description{
Use a codes table to recode variable. Might be useful for translation of labels, for example.
}
\examples{
library(readxl)
codes <- read_excel("countryCodes.xls", 1)
replace_by_table(mydata$var1, codes, "iso2c", "full.country.names")

}
