% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.r
\name{reverse}
\alias{reverse}
\title{Reverse values and value labels}
\usage{
reverse(var, preserve.labels = T)
}
\arguments{
\item{var}{variable to reverse.}

\item{preserve.labels}{Attempt saving labels.}
}
\description{
Returns same variable with reversed values. Reversing makes use of observed values rather than stored factor levels or attributes.
}
