% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_helpers.R
\name{search_random}
\alias{search_random}
\title{Find which random effects could be added to a lmer model}
\usage{
search_random(lmerfit, terms = NA, boot = F)
}
\arguments{
\item{lmerfit}{lmer fit}

\item{terms}{terms, if NA (default) uses each variable from fixed effects}

\item{boot}{Apply bootstrap?}
}
\description{
Find which random effects could be added to a lmer model
}
