% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_helpers.R
\name{sem_tab}
\alias{sem_tab}
\title{SEM tab}
\usage{
sem_tab(..., what = c("loadings", "intercepts", "slopes"), std = F)
}
\arguments{
\item{...}{Fitted lavaan models as separate arguments or a list of models}

\item{what}{Kinds of parameters to report. Possible values "loadings" and "intercepts"}
}
\description{
SEM tab
}
