% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{sig_seg}
\alias{sig_seg}
\title{Prepares data for significance staples on ggplot}
\usage{
sig_seg(df, compare.var, y.var, group.var, heights, ends)
}
\arguments{
\item{df}{Aggregated data frame}

\item{compare.var}{Indicator of categories compared}

\item{y.var}{Means or other compared quantities (used to define the ends of the staples)}

\item{group.var}{Any additional variable in the dataset (e.g., groupuing variable)}

\item{heights}{Numeric vector. Height or width of each staple.}

\item{ends}{Numeric vector. Fixed ends for even staple graphs, in case specified, `y.var` is ignored.}
}
\description{
Prepares data for significance staples on ggplot
}
