% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sig_to_bold}
\alias{sig_to_bold}
\title{Conditional formatting}
\usage{
sig_to_bold(
  x,
  star = F,
  bold.thresh = 1,
  neg.red = T,
  small.thresh = 0.01,
  replace.small = NA,
  extra.style = ""
)
}
\arguments{
\item{x}{Character vector}

\item{star}{Removes stars if FALSE.}

\item{bold.thresh}{Bold font if the absolute value is larger than this threshold.}

\item{neg.red}{Red font if the value is negative value.}

\item{small.thresh}{Replaces the value with `<|.01|` if the magnitude is smaller than the indicated number. Default is .01.}

\item{replace.small}{What to replace small values with. Default is `<|.01|`}

\item{extra.style}{Add extra CSS style to the span tag.}
}
\description{
Adds formatting to a vector of parameters
}
\details{
Can be used within `mutate()` and passed to `knitr::kable` with parameter `kable.options=list(escape=FALSE)`
}
