% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MplusHelpers.R
\name{tech11.14}
\alias{tech11.14}
\title{Extracts Tech11 and Tech14 (Tests for mixture models)}
\usage{
tech11.14(output)
}
\arguments{
\item{output}{Output either from `readModels()$output` or the raw file read by e.g. `readLines(Mplus output)`}
}
\value{
A list of data.frames with tests.
}
\description{
Extracts Tech11 and Tech14 containing VLMR, adjusted LMR, as well as BLRT ests for mixture models (L-1 vs L number of components)
}
