% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_book.R
\name{unhaven}
\alias{unhaven}
\title{Drop the structure added by haven package}
\usage{
unhaven(haven.df)
}
\arguments{
\item{haven.df}{Data.frame or tibble with columns of class haven_labelled. If the class of the variables is different, they will be just copied as they are. Tibble (list-like) structure is also removed if present.}
}
\description{
Drop the structure added by haven package
}
