% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{verb}
\alias{verb}
\title{Convenience function for logging the code}
\usage{
verb(...)
}
\arguments{
\item{...}{Character strings.}
}
\description{
Convenience function for logging the code
}
\details{
Prints text and objects in order to log and debug the code.
}
