% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_2.R
\name{view_2}
\alias{view_2}
\title{Data Table Alternative Viewer}
\usage{
view_2(df, filename = "dt_tmp.html", size = 100)
}
\arguments{
\item{df}{Data frame to view.}

\item{size}{Number of randomly selected observations, defaults to 100.}

\item{path}{Path where to save the temporary file, defaults to "dt_tmp.html".}
}
\value{
Nothing, opens browser window
}
\description{
This is a custom function to view data frames with many variables, and include variable names, as defined in the labelled package, if they exist.
}
\examples{
view_2(cars)
View = view_2 # replaces RStudio standard viewer
}
