% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_helpers.R
\name{vif_mer}
\alias{vif_mer}
\title{Compute variance inflation factor for mer objects}
\usage{
vif_mer(fit)
}
\arguments{
\item{fit}{Object of class lmer.}
}
\description{
Compute variance inflation factor for mer objects
}
\details{
adapted from rms::vif
}
