% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_helpers.R
\name{add_term}
\alias{add_term}
\title{Add or remove terms to a lmer formula string}
\usage{
add_term(x, fixed = NULL, random = NULL, drop = NULL, drop.random = NULL)
}
\arguments{
\item{x}{Either mer formula object, string containing formula, lmer object. If latter is the case, the updated object will be returned instead of a new formula.}

\item{fixed}{List of character strings containing variable names or terms to include in the formula as fixed terms.}

\item{random}{List of character strings containing variable names or terms to include in the formula as random terms.}

\item{drop}{List of character strings containing variable names or fixed terms to exclude from the formula.}

\item{drop.random}{List of character strings containing variable names or random terms to exclude from the formula.}
}
\description{
Add or remove terms to a lmer formula string
}
\examples{
f1 <- "a ~ b + c + (1| group)"
add_term(f1, , "Conservation")
add_term(f1, "Conservation*GDP")
add_term(f1, "Conservation*GDP", "Conservation")

add_term(f1, drop="female", drop.random = "g.agea")

}
