% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_helpers.R
\name{aggr_and_merge}
\alias{aggr_and_merge}
\title{Creates group-level variables by aggregating individual-level variables and adding it to the df}
\usage{
aggr_and_merge(
  variables,
  group,
  data,
  FUN = "mean",
  prefix = paste0(group, ".")
)
}
\arguments{
\item{variables}{Variable to be aggregated.}

\item{group}{Grouping variable.}

\item{data}{A dataset with individual-level data.}

\item{FUN}{Aggregation function, mean by default.}

\item{prefix}{Character string, what should be added to the  names of aggregated variables.}
}
\value{
Returns the original data frame with new group-level variables appended.
}
\description{
Creates group-level variables by aggregating individual-level variables and adding it to the df
}
\details{
Alternative to `dplyr::group_by` + `dplyr::mutate`. Can be used within functions.
}
