% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_helpers.R
\name{cor_between}
\alias{cor_between}
\title{Correlation between group aggregates}
\usage{
cor_between(var1, var2, group, data, print = T, ...)
}
\arguments{
\item{var1}{Character name of variable 1, or a list of characters (in this case var2 is ignored)}

\item{var2}{Character name of variable 2.}

\item{group}{Character name of group variable.}

\item{data}{Dataset}

\item{...}{Passed to `cor.test`}
}
\description{
Correlation between group aggregates
}
