% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{cor_table}
\alias{cor_table}
\title{Correlation table with stars}
\usage{
cor_table(d, method = "spearman", star = TRUE)
}
\arguments{
\item{d}{Data.frame}

\item{method}{Method from `cor` function. spearman by default.}
}
\description{
Produces data frame of correlations with stars standing for significance
}
