% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_viewer.R
\name{df_to_viewer}
\alias{df_to_viewer}
\alias{to_viewer}
\title{Shows data.frame in RStudio viewer}
\usage{
df_to_viewer(
  x,
  rownames = TRUE,
  summ = F,
  kable = FALSE,
  by = NULL,
  html = FALSE,
  colformat = NULL,
  sg.style = "ajs",
  k.style = "default",
  digits = 2,
  ...
)
}
\arguments{
\item{rownames}{Logical, if rownames should be shown.}

\item{summ}{Passed to `summary` argument of stargazer.}

\item{kable}{logical, if data.frame should be passed to kable. If true, ... will be passed to function kableExtra::kable_styling}

\item{by}{The name of variable to group the rows by. Used only when `kable=TRUE`}

\item{html}{If viewer should be ignored and just html code returned (useful to embed in Rmarkdown outputs}

\item{colformat}{Formats each column using `sprintf`. If numeric, applies rounding only, if character, is passed to `sprintf` directly. Can be named to match x's column names. If not named, must have the same length as there are columns in x.}

\item{sg.style}{'style' argument of `stargazer`.}

\item{k.style}{Applies custom kableStyling style, might take two values "default", and "custom".}

\item{...}{Other arguments passed to `stargazer` or `kableStyling()`.}

\item{df}{data.frame to show. Also works with a list of fitted lm() models.}
}
\description{
via stargazer
}
