% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_viewer.R
\name{df_to_viewer}
\alias{df_to_viewer}
\title{Shows data.frame in RStudio viewer}
\usage{
df_to_viewer(df, rownames = TRUE, ...)
}
\arguments{
\item{df}{data.frame to show. Also works with a list of fitted lm() models.}

\item{rownames}{Logical, if rownames should be shown.}

\item{...}{Other arguments passed to stargazer.}
}
\description{
via stargazer
}
