% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_viewer.R
\name{df_to_viewer}
\alias{df_to_viewer}
\alias{to_viewer}
\title{Shows data.frame or model(s) in RStudio viewer}
\usage{
df_to_viewer(
  x,
  rownames = TRUE,
  summ = F,
  kable = TRUE,
  by = NULL,
  html = FALSE,
  colformat = NULL,
  sg.style = "ajs",
  k.style = "default",
  digits = 2,
  kable.options = list(),
  cells.by.merge = T,
  cells.by.indent = T,
  ...
)
}
\arguments{
\item{rownames}{Logical, if rownames should be shown.}

\item{summ}{Passed to `summary` argument of stargazer.}

\item{kable}{logical. If data.frame should be passed to kable.  If true, ... will be passed to function kableExtra::kable_styling. Defaults to TRUE. If FALSE, passes df to stargazer.}

\item{by}{The name of variable to group the rows by. Used only when `kable=TRUE`}

\item{html}{If viewer should be ignored and just html code returned (useful to embed in Rmarkdown outputs}

\item{colformat}{Formats each column using `sprintf`. If numeric, applies rounding only, if character, is passed to `sprintf` directly. Can be named to match x's column names. If not named, must have the same length as there are columns in x.}

\item{sg.style}{'style' argument of `stargazer`.}

\item{k.style}{Applies custom kableStyling style, might take two values "default", and "custom".}

\item{kable.options}{A list of arguments passed to `kable` function.}

\item{cells.by.merge}{The headers created by `by` argument can be a single cell spanning all the columns (TRUE) or the subheader can be put in the new line's first column (FALSE). The latter tables are easier to hande in, e.g. MS Word.}

\item{cells.by.indent}{Logical. Whether 'by' headers should have an indent.}

\item{...}{Other arguments passed to `stargazer` or `kableStyling()`.}

\item{df}{data.frame to show. Also works with a list of fitted models supported by .}
}
\description{
via stargazer and knitr
}
