% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{eststar}
\alias{eststar}
\title{Quickly generate string from estimates and p-values}
\usage{
eststar(est, pvalue, digits = 2, na = "", ci.low = NA, ci.hi = NA)
}
\arguments{
\item{est}{estimate}

\item{pvalue}{p-value}

\item{digits}{round to this number of decimals}

\item{na}{return this if `est` is `NA`}
}
\description{
Quickly generate string from estimates and p-values
}
