% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MplusHelpers.R
\name{f}
\alias{f}
\title{Format the number quickly}
\usage{
f(x, digits = 3)
}
\arguments{
\item{x}{any numeric vector or value.}

\item{digits}{number of decimals.}
}
\value{
Character of the formatted
}
\description{
Format the number quickly
}
