% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{get_net}
\alias{get_net}
\title{Make a quick network graph}
\usage{
get_net(
  df = NULL,
  cv = NULL,
  n = NULL,
  title = "",
  legend = "none",
  gamma = 0.3,
  layout = "fr",
  start = "glasso",
  stepwise = T,
  seed = 1,
  group.id = NA,
  item_label = function(x) gsub("\\\\D", "", x),
  return.df = F,
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{cv}{Correlation matrix (if df is NULL)}

\item{title}{Title of the plot}

\item{legend}{Legend position}

\item{gamma}{Gamma parameter for the model selection function, see `ggmModSelect`}

\item{layout}{Default is "fr". Network layout. See more at `ggraph::ggraph`}

\item{start}{Default is "glasso", see `ggmModSelect`}

\item{stepwise}{Default is TRUE, see `ggmModSelect`}

\item{seed}{Attempt to make plots replicable.}

\item{group.id}{Vector of regex expressions to apply to variable names and use it to group variables.}

\item{item_label}{Function that converts the item names to node labels.}

\item{return.df}{Logical, if true, returns tidygraph dataset instead of the plot.}
}
\description{
Make a quick network graph
}
\examples{
data("HolzingerSwineford1939", package = "lavaan")
get_net(HolzingerSwineford1939[, 7:15],
       group.id = c(first.factor = "1|2|3",
                    second = "4|5|6",
                    third = "7|8|9"),
       legend = "bottom"
)

}
