% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_helpers.R
\name{good_table}
\alias{good_table}
\title{Function producing a good table for many ML regressions}
\usage{
<<<<<<< HEAD
good_table(models, fit.stats = c("ICC", "random", "random.p", "R2",
  "fit", "LRT", "REML", "VIF"), mod.names = "", show.viewer = TRUE,
  ...)
=======
good_table(models, fit.stats = c("ICC", "random", "random.p", "R2", "fit",
  "LRT", "REML", "VIF"), mod.names = "", show.viewer = TRUE, ...)
>>>>>>> 3ba7ee8f1e8c681e900d55fc66ae9042bcc8c4f5
}
\arguments{
\item{models}{List of the fitted lmer objects}

\item{fit.stats}{What fit statistics to compute/extract and show? Possible options are "ICC", "random", "R2s", "fit", "LRT", "REML", "VIF". See "Details".}

\item{mod.names}{Vector of the same length as models list, giving names to each model.}

\item{show.viewer}{Logical. Whether the resulting table should be shown in the RStudio viewer. If FALSE then the file "good_table_output.html" is saved to working directory.}

\item{...}{Arguments passed to stargazer.}
}
\description{
Function producing a good table for many ML regressions
}
\details{
<<<<<<< HEAD
The list of possible fit.stats options:
=======
>>>>>>> 3ba7ee8f1e8c681e900d55fc66ae9042bcc8c4f5
\describe{
\item{fit}{Shows deviance (-2*logLikelihood), AIC, BIC, number of parameters, number of groups, number of observations, if the model converged}
\item{ICC}{Computes intra-class correlation by fitting an empty model and computing a ratio of first-level and intercept variances.   }
\item{R2}{Computes R-square by fitting an empoty model and computing a ratio of residuals in an empty model and in the current model. }
\item{LRT}{Computes Likelihood ratio test; all the models should be fitted to the same sample and be in order of nestedness, otherwise the test fails. }
\item{random}{Adds variances of all the random effects.}
\item{random.p}{Adds variances of all the random effects and implements bootstrapping (\code{lme4::confint.merMod}) in order to get confidence intervals and this p.values for variances of random effects. }
\item{REML}{Shows if REML was used to fit the model.}
}
}
\note{
Issues to implement:
\itemize{
\item Might be veeery slow (something to work on)
<<<<<<< HEAD
}
}
\examples{
\donttest{
m1<- lmer(HE ~ age + gender + (1|country), ess.data.d, weights=dweight)
m2<- lmer(HE ~ age + gender + (1+age + gender|country), ess.data.d, weights=dweight)
good_table(list(m1, m2))

# Another option
f1 <- as.formula(HE ~ age + gender + (1|country))
f2 <- as.formula(HE ~ age + gender + (1|country))
good_table(lapply(c(f1, f2), function(x) lmer(x, ess.data.d, weights=dweight)))
}

=======
\item Add row indication of non-convergence
}
}
\examples{
#m1<- lmer(HE ~ age + gender + (1|country), ess.data.d, weights=dweight)
#m2<- lmer(HE ~ age + gender + (1+age + gender|country), ess.data.d, weights=dweight)
#good_table(list(m1, m2))

# Another option
#f1 <- as.formula(HE ~ age + gender + (1|country))
#f2 <- as.formula(HE ~ age + gender + (1|country))
#good_table(lapply(c(f1, f2), function(x) lmer(x, ess.data.d, weights=dweight)))
>>>>>>> 3ba7ee8f1e8c681e900d55fc66ae9042bcc8c4f5

}
