% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_helpers.R
\name{group_center}
\alias{group_center}
\title{Group-centering of one or more variables}
\usage{
group_center(variables, group, data, prefix = "g.")
}
\description{
Group-centering of one or more variables
}
