% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_helpers.R
\name{lav_compare}
\alias{lav_compare}
\title{Combine fit measures form several models and compare}
\usage{
lav_compare(
  ...,
  what = c("cfi", "tli", "rmsea", "srmr", "bic", "df"),
  LRT = F,
  print = T
)
}
\arguments{
\item{...}{lavaan fitted objects}

\item{what}{Character vector of fit indices as given in \pkg{lavaan} `fitMeasures()`}

\item{LRT}{Logical. Whether to include likelihood ratio test (anova) comparing the models.}

\item{print}{Logical. Whether to print the results to the console.}
}
\description{
Combine fit measures form several models and compare
}
\details{
Extracts select fit measures from lavaan objects and lists it in the tables.
}
\examples{
data("HolzingerSwineford1939", package="lavaan")
m1 = cfa("F1 =~ x1 + x2 + x3 + x4 + x5", estimator= "mlr", HolzingerSwineford1939)
m2 = cfa("F1 =~ x1 + x2 + x3 + x4 + x5; x4 ~~ x5", estimator= "mlr", HolzingerSwineford1939)
lav_compare(m1, m2, what = c("cfi.scaled", "rmsea.scaled") )


}
