% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_helpers.R
\name{lav_to_graph}
\alias{lav_to_graph}
\title{Create a path diagram using lavaan syntax}
\usage{
lav_to_graph(m, layout = "dot", adds = NULL, file = NA, ...)
}
\arguments{
\item{m}{lavaan syntax model, e.g. "F =~ a1 + a2 + a3 + a4; a1 ~~ a2; d ~ F "}

\item{file}{File to save svg code, usually with 'svg' extension.}

\item{...}{arguments passed of DiagrammeR::grViz function.}
}
\description{
Converts lavaan syntax to graphviz code. Currently works with simple models only. Requires 'DiagrammeR' package.
}
\examples{
lav_to_graph("F =~ a1 + a2 + a3 + a4")
lav_to_graph("F =~ a1 + a2 + a3 + a4; a1 ~~ a2; d ~ F ", engine = "neato")

}
