% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MplusHelpers.R
\name{partable_mplus}
\alias{partable_mplus}
\title{Summarize parameters from several Mplus models read by 'readModels'}
\usage{
partable_mplus(models, std = FALSE, se = T, stars = T, digits = 2)
}
\arguments{
\item{models}{A list of models read by MplusAutomation::readModels}

\item{std}{If stdyx standardized coefs should be reported.}

\item{se}{If SEs should be included.}

\item{stars}{Logical. if significance stars should be added.}

\item{digits}{Numeric value, how many decimals required.}
}
\value{
Returns a single data frame with unstandardized parameters.
}
\description{
Creates a table (dataframe) of parameters extacted from Mplus
}
