% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.r
\name{reverse}
\alias{reverse}
\title{Reverse values and value labels}
\usage{
reverse(var)
}
\arguments{
\item{var}{variable to reverse.}
}
\description{
I don't know why it wasn't developed before. Returns same variable with reversed values.
}
\details{
Reversing makes use of observed values rather than stored factor levels or attributes.
}
