% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{scatter_means_ci}
\alias{scatter_means_ci}
\title{Computes means by group and plots them in as a scatterplot}
\usage{
scatter_means_ci(
  variable1,
  variable2,
  group,
  data,
  plot = TRUE,
  print = TRUE,
  smooth.method = "lm",
  drop.groups = NULL
)
}
\arguments{
\item{variable1}{Variable to plot on x axis}

\item{variable2}{Variable to plot on y axis}

\item{group}{Grouping variable}

\item{data}{Data frame}

\item{plot}{Logical. Should the plot be created?}

\item{print}{Logical. Should the result be printed in the console?}

\item{smooth.method}{Method for smoothing line. "lm" by default.}

\item{drop.groups}{Vector of groups to be excluded from the analysis.}
}
\description{
Computes means by group and plots them in as a scatterplot
}
\details{
Recommended to use with `with` function, as in example.
}
\examples{
scatter_means_ci('alcfreq', 'HE', 'cntry', ESS, plot=T, print=TRUE)

}
