% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{scatter_means_ci}
\alias{scatter_means_ci}
\title{Computes means by group and plots them in as a scatterplot}
\usage{
scatter_means_ci(var1, var2, group, plot = TRUE, print = TRUE)
}
\arguments{
\item{var1}{Variable to plot on x axis}

\item{var2}{Variable to plot on y axis}

\item{group}{Grouping variable}

\item{plot}{Logical. Should the plot be created?}

\item{print}{Logical. Should the result be printed in the console?}
}
\description{
Computes means by group and plots them in as a scatterplot
}
\details{
Recommended to use with `with` function, as in example.
}
\examples{
 with(ess.data.d, scatter_means_ci(alcfreq_r, HE, cntry, plot=T, print=TRUE))
}
