% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_helpers.R
\name{sem_tab}
\alias{sem_tab}
\title{Table with parameters from a lavaan model}
\usage{
sem_tab(
  ...,
  what = c("loadings", "intercepts", "slopes", "covariances", "others"),
  std = F,
  ci = F
)
}
\arguments{
\item{...}{Fitted lavaan models as separate arguments or a list of models}

\item{what}{Kinds of parameters to report. Possible values "loadings", "intercepts",  "slopes", "covariances", and "others"}

\item{std}{Whether to report standardized or unstandardized estimates}

\item{ci}{Whether to include confidence intervals}
}
\description{
Table with parameters from a lavaan model
}
