% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{stacked_bar}
\alias{stacked_bar}
\title{Quick stacked bar plot}
\usage{
stacked_bar(
  variable,
  group,
  sort.cat = 1,
  colors = NA,
  include.na = FALSE,
  labs = TRUE,
  label.col = NA,
  format.label = "2\%",
  label_col = NA,
  weight = NA,
  flip = F,
  cat.rev = F
)
}
\arguments{
\item{variable}{Variable to be plotted.}

\item{group}{Grouping variable to cross-tabulate and compute.}

\item{sort.cat}{Category to sort the stacked chart.}

\item{colors}{Vector of colors distinguishing groups.}

\item{include.na}{Logical. Should NAs be a separate level or should they be excluded? FALSE by default.}

\item{labs}{Logical. Should percents labels be added to the plot. TRUE by default.}

\item{label.col}{In case labels are added, vector of labels colors.}

\item{format.label}{Character of one digit standing for decimal places and one symbol added in the end, for example "1\%" would result in "99.9\%" whereas "2$" would lead to "99.99$".}

\item{label_col}{Deprecated, use `label.col` instead.}

\item{weight}{If there is a survey weight, variable to use in crosstab computation.}

\item{flip}{Logical, whether to flip the coordinates. FALSE by default.}

\item{cat.rev}{Logical, whether to reverse the order of categories in the legend and in the bars. FALSE by default.}
}
\description{
Quick stacked bar plot
}
\details{
Computes proportions cross-table and plots it in a nice way. ggplot-based, so `+theme()` and other arguments can be added directly to transform the plot.
}
