% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_values.R
\docType{data}
\name{values}
\alias{values}
\title{List of Schwartz values}
\format{An object of class \code{list} of length 9.}
\usage{
values
}
\description{
List of Schwartz values
}
\author{
Collected using multiple publications by S.H.Schwartz and coauthors.
}
\keyword{datasets}
